Add basic calculator GUI
This commit implements the StandardCalculatorView user control, which is used
in the basic calculator mode that doesn't offer expressions nor goniometric
functions.

There's a custom control called DoubleButton, which encapsulates two buttons
into a single control with no apparent visual division. It's used for packing
together pairs of operations, which is done to make the calculator look more
compact.

The StandardCalculatorViewModel class will be used as the main functional entity
of this view and will be implemented in following revisions.

I've chosen the Roboto font, which is now included in the app's resources. This
may or may not be subject to further discussion. Also, a hopefully temporary
app icon has been added.