Implement extended calculator mode
The extended calculator mode uses the maths service expression solver
instead of using maths service methods directly. The buttons merely append
numbers or operators to the display and when the Equals button is pressed,
the viewmodel converts the expression to a correct format (replacing the UTF
times and divide chars with an asterisk and a slash).

The biggest issue to face here was keeping track of the caret position
to append or remove the numbers to/from a correct place on the display.
The Avalonia's TextBox discards the caret position when the control loses
focus. For this reason, a DisplayTextBox class has been added that extends
the default TextBox and overrides the OnLostFocus method where it saves
the caret index and then restores it.