Change what kind of packages publishing make invokes
For the Linux installer, we are going to use a runtime-dependent single-file executable.
After this change, make will produce runtime-depended binaries, executables for
both Windows and Linux and self-contained executables (that contain the runtime) for both.
What means of distribution are we going to use on Windows is still not decided.